spokenly-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
deepgram-sdk>=4.0.0
firebase-admin==6.2.0
python-multipart==0.0.6
aiohttp==3.10.10
